daemon: avoid infinite crash loops
Export an environment variable as a mark for abrtd (abrt-server) to
identify processes directly involved in ABRT post-mortem processing.

We must not run post-mortem EVENTs on problem directories caused by ABRT
itself because we could create an infinite loop.

There are to ways how to handle such directories:
    * in non-debug mode: log a short message and remove them without
      other actions - we must not leave them in the dump location by
      default because the dump location would be growing

    * in debug mode: log a more verbose message and leave them as they
      are - we don need to have worries about the dump location growing
      because someone intentionally enable the debug mode

Related: rhbz#1246539

Signed-off-by: Jakub Filak <jfilak@redhat.com>