ccpp: avoid running elfutils under root
This commit improves security of abrt-hook-ccpp when generating
core_backtrace from a dying process using elfutils functions. elfutils
is a complex library, therefore running those functions on a random
process under root user is security risky.

This patch teaches abrt-hook-ccpp to change all process-related user
ID's and group ID's to UID resp. GID of the process. This is quite
tricky when the process' UID does not equal effective UID (resp. GID and
effective GID) - IOW a set-UID (set-GID) process. In such a cause,
abrt-hook-ccpp checks UID and effective UID and uses the one which is
not 0.

Nevertheless, it is not possible to run all elfutils functions under an
unprivileged user because elfutils needs to read the /proc/[pid]/auxv,
/proc/[pid]/maps and /proc/[pid]/mem file that are not readable by other
users. The mem file is required when the process' executable is removed,
so in such a case core_backtrace generation will fail.

This commit also teaches abrt-hook-ccpp to drop all capabilities before
calling elfutils because no capabilities should be needed and we should
try to reduce effects of a possible attack disaster.

Signed-off-by: Adam Sulc <xmineral@seznam.cz>
Signed-off-by: Jakub Filak <jfilak@redhat.com>