daemon: trigger dump location cleanup after detection
This commit restores the old behaviour where the cleanup algorithm was
started right after new dump directory is created. This prevents piling
up of new dump directories which could lead to consumption of too much
disk space. The piling up of dump directories is currently prevented by
the plugins removing old dump directories on their own, which is in fact
problematic because the plugins don't know about each other and that causes
race conditions.

The post-create EVENT execution was moved from abrtd to abrt-server in
commit b6640620e27a029b3f1f8dcec22fb4c95e48db2a in order to replace the
inotify watch in abrtd with the /creation_notification method of
abrt-server.

What are the cases we must deal with
-----------------------------------

1) an old directory is to be removed
2) one of the queued directory is to be removed
3) currently processing directory is to be removed

The case 1) is not problematic at all (except removing directories that
are currently being handled by users).

The case 2) would cause an error message produced by abrt-handle-event
waked up from waiting for post-create.lock - the error message could be
avoided by ignoring the error in case of running post-create EVENT.

The case 3) is extremely problematic and must be avoid in all situation.
There is no other way how to avoid this case without a central
synchronization algorithm. One could claim that we should lock the
currently processed dump directory and don't removed the locked ones but
libreport's locking algorithm doesn't support recursive locking between
processes - however, the recursive inter process locking would get rid
of the case 1). Or abrt-handle-event could write the handled directory
name to a new file but it is not clear where the file would be consumed
as there is no authority doing the cleanup. And, what is the worst,
communication trough files will lead to another type race conditions.

What this patch introduces
--------------------------

This patch adds communication between abrtd and its child processes
abrt-server. When abrt-server is asked to run post-create EVENT, it
sends the "NEW_PROBLEM_DETECTED: $DUMP_DIR" message to abrtd over
STDERR. STDERR is used because STDOUT is occupied by the socket (we
might want to make it less obfuscated in future and use a FIFO
or something else, but now I am happy with using STDERR). abrtd
then pushes the abrt-server process to a queue used to track abrt-server
processes wanting to run post-create EVENT. When a process from the
queue is to be executed abrtd sends it SIGUSR1 signal. If a dump
directory of any of queued process was removed, abrtd sends the relevant
abrt-server process SIGINT signal.

Signed-off-by: Jakub Filak <jfilak@redhat.com>