testsuite: do not die if crash is not generated
get_crash_path() died in case the crash is not generated. If some test
changes abrt's configuration in the beginning of the test and
get_crash_path() dies because of the crash is not generated, the test
ends without changing the configuration back. This can cause failing of
other tests.

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>