lib: normalize slashes of configured paths
ABRT tools often compare file system paths. This is stemming from the
nature of ABRT data that are regular directories and files.

To make those compassions reliable one must compare canonical paths.

This patch adds normalization/canonicalization of slashes of paths
defined in configuration files.

Please notice that we cannot use the realpath function because it works
only for existing files and directories and, for example, paths of the
DumpLocation option may not exist.

I intentionally ignore "." and ".." because I do not expect users to put
such paths to configuration file and I do not want to make the
processing too complex. On the other hand, slashes can be multiplied by
accident or by a script.

Signed-off-by: Jakub Filak <jfilak@redhat.com>