daemon: send  base names from abrt-server to abrtd
The /creation_notification method accepts arbitrary file system path and
it might cause abrtd some troubles if the path is not in canonical form.

ABRT enforces the notified paths to be directories in /var/spool/abrt
but ignores number of slash occurrences as it is not security relevant.

Switching to base names will assure that the problems associated with file
system path forms will be gone and no new problems will be introduced.

The other option is to use realpath() but this function resolves
symbolic links and if /var/spool is a symbolic link, we will be in
trouble.

Related: rhbz#1132459

Signed-off-by: Jakub Filak <jfilak@redhat.com>