a-a-ureport: print out exit codes in verbose mode
Exit code of subcommands are rather useless to regular users.

The message removed by this commit was there just to make sure the
action will not exit without any output. However, if the action is
executed from an event, libreport prints out the following message:
('report_uReport' exited with 1)

If the action is execute manually from command line, then we can assume
the user who issued the command can run it again with increased
verbosity.

Signed-off-by: Jakub Filak <jfilak@redhat.com>