a-a-analyze-c: create crash_function element
Create a crash_function element from core_backtrace in post-create
event. The crash_function element was only created by
abrt-action-analyze-backtrace tool and was not run automatically after
crash detecting.

We need creation the element before notify event is run because
reporter-systemd-journal is run in notify event and it uses
the crash_function element.

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>