ccpp: use abort() to exit in debug mode
libreport tends to exit on failures and if the error message isn't good
enough, you won't be able to find the line which caused the exit.

This commit teaches abrt-hook-ccpp to use the 'abort' mode. The reason
behind selecting 200 as the appropriate levels is that I believe that
this is more 'hardcore' than enabling core dumping of abrt-hook-ccpp
(don't forget that abort() will cause abrt-hook-ccpp to drop a core file).

The mode can be set via ENVIRONMENT variable but there is no way how to
specify the variable for abrt-hook-ccpp executed by kernel from
/proc/sys/kernel/core_pattern.

Signed-off-by: Jakub Filak <jfilak@redhat.com>