ccpp: fast dumping and abrt core limit
This commit introduces a new configuration option MaxCoreFileSize. The
option will be compared to MaxCrashReportsSize and the minimum will be
used as the limit for the size of the core dump file in a dump directory.

This commit replaces the read once write twice algorithm with a tee + splice
algorithm. tee() does zero copy - it should just increment counters -
and splice() moves bytes between file descriptors without the need to
copy them to user space. Basically the new algorithm is the same but
without the need to copy data to user space. However, the original
algorithm was testing the buffer for 0s and if the entire buffer was 0,
then lseek() was performed instead of write(). The 0 check was there to
make the dumping faster and this is no longer needed as the tee +
splice is faster than the original read once write twice algorithm.

Signed-off-by: Jakub Filak <jfilak@redhat.com>