daemon: properly shutdown socket connection
Upon a new connection from a peer abrtd creates a connection fd and
spawns a new process that executes abrt-server. The new process
duplicates the connection fd to STDIN_FILENO and STDOUT_FILENO. When
abrt-sever is done with creation of a dump directory it sends a message
about that and closes STDOUT_FILENO to shut down the connection.

However, abrt-server leaves STDIN_FILENO opened and thus the connection
is not closed. This has a negative effect on the Python exception handler
that waits until the connection is closed - effectively, it delays
restarts of crashed processes.

Signed-off-by: Jakub Filak <jfilak@redhat.com>