ccpp: retain partial core_backtrace upon error
The idea behind removing core_backtrace if the core-backtrace generator
fails was to remove strange files and let ABRT post-mortem processing to
create the item properly.

It was a nice idea but we should do better. We must remove only empty
file because in other cases core_backtrace can be the only evidence of
the crash - for example if SaveFullCore is disabled.

Signed-off-by: Jakub Filak <jfilak@redhat.com>