koops_event: add check to restrict reporting of MCEs
Added creation of new element 'mce' for oopses, where it's possible values
are 'fatal' or 'non-fatal'.

Added restriction on reporting MCEs, we don't want to send uReports for
both fatal and non-fatal MCEs since they don't contain backtrace and
so are invalid uReports.
Reporting using GUI of non-fatal MCEs is disabled.

In future it will be nice to implement specialized MCE watcher tool.

Resolves abrt/abrt#1033