changing load location from bin to libexec
abrt-action-save-container-data was located in bin folder and it doesn't have a man page.
We investigated that it is not necessary for abrt-action-save-container-data to have man page so we moved it from bit folder to libexec folder.
In code we had to change lines for abrt-action-save-container-data which was connected to bin folder to libexec folder.