a-h-event: Do not deduplicate different containers
When the same problem happens in two different containers, it gets
recognized as duplicate and second occurrence is lost. Which means, that
we do not know, that the problem happened in second container.

This commit fixes that. Two problems cannot be duplicate, when they are
not from the same container.

For better understanding here goes a real use case:

I, as a devops, test my App in containers. Testing is sometimes interrupted and
does not finish. It happened again so I use Cockpit to quickly check about the
container. I don't see there anything unusual. (Because crash from this
        container happened earlier and is only displayed in some quite
        old (and maybe not even existing) container)
However, after this commit I see in Cockpit exclamation mark next to
containers, in which crash happened and can see detailed information about this
problem.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>