spec: fix ambiguous Python 2 dependency declarations
This package uses names with ambiguous python- prefix in requirements.

According to Fedora Packaging guidelines for Python, packages must use
names with either python2- or python3- prefix in requirements where
available.
We are aiming to rename python-* dependencies to python2-*, so we can
later switch the python-* namespace to Python 3.

This PR is part of Fedora's Switch to Python 3 effort.

Note that, although this PR was created automatically, we will respond
to any comments or issues which you might find with it. We will keep the
PR open for review for a week, and if there's no feedback we'll merge
it.

Koji scratch build:
https://koji.fedoraproject.org/koji/taskinfo?taskID=23642779
Note: please do not backport this to f26 branch(es) as some of the
modified requirements are not available there

Special thanks to Iryna Shcherbina <ishcherb@redhat.com> who did the
patch.

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>