applet: Port to GApplication/GNotification
This commit ports the applet from its current use of libnotify to
GNotification and, consequently, GApplication. That nets us persistent
notifications following the removal of the tray icon, since currently
they are lost when the applet goes down. Additionally, this makes the
applet D-Bus activatable for when the notifications are clicked, so that
the applet itself is respawned and gnome-abrt is launched. A bit
roundabout, but that’s that.

Closes https://github.com/abrt/abrt/issues/1085

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>