testsuite: abrt-action-ureport: Fix fakefaf.py
Some fallout from 74966d9eb09e1e29cfdba642e44d965a20494e4c, responses
are written as strings, but the file stream is binary. Additionally, the
default log_message() handler does not flush stderr, resulting in empty
logs when we redirect streams to a file and run in the background.

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>