plugins: abrt-action-install-debuginfo: Catch BrokenPipeError
This commit adds a global handler for BrokenPipeError to account for
cases, where reporting is canceled from report-gtk. Since the script
usually runs under a different user, terminating the process will not
work as expected, and it might end up writing to a closed stdout pipe.

Using a class to stash command-line options is an implementation detail,
as added scoping results in UnboundLocalErrors due to assignments.
Marking those variables as global would have been the alternative, but
blegh.

https://bugzilla.redhat.com/show_bug.cgi?id=1501178

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>