doc: Update man pages and comments in config files
Update descriptions of options in configuration files, in their man
pages and man pages of the tools that use them.

Main changes include a formatting reform, wording improvements and
updates to reflect real defaults.

All the configuration options are now commented out with the
authoritative defaults being hardcoded and mentioned in the comments and
documentation.