cli: Print usage if option passed with no command
Given that the help string for --authenticate is misleading, one might
try to run `abrt --authenticate` and expect it to work as the standalone
version (with implicit list command), however, it should rather be a
command option, since it modifies the behavior of a command being
executed and the said command might not even do anything that could be
authenticated.

The most logical fix is to require a command argument if any non-command
option is passed (aside from action options, e.g. help and version).

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1770034

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>