koops: Avoid explicit dependency on hash size
Change the API so that koops_hash_str and koops_hash_str_ext return the
heap-allocated encoded digest instead of taking a pointer and returning
a status code.

This eliminates the need to know the size of the resulting hash in
advance, which should solve the build problems following recent
libreport changes.