plugins: abrt-action-analyze-oops: Refactor code
This commit GLib-ifies the code and fixes the return code in case of
success, as currently NULL != hash_str is returned, which will be the
opposite of what it should be
(caused by 5284fd63f1d2d9da09fd1793e49ab734557ea046, my bad).

This is one step towards getting rid of the libreport command-line
parsing nonsense.