plugins: journal: Fix ci_mapping being overwritten
Builds on top of 98c4205a717dbca34c3a889c6c133ed06feb335f, which was a
horrible mess. In systems that use the LP64 data model (Linux), writing
to ci_pid in struct crash_info through a long pointer would result in
ci_mapping being overwritten, leading to crashes when collecting core
dumps from the journal. This commit address the original issue in a
different - several wrappers are provided for a macro that casts the
result from strtol() to a suitable type. The potential issue with
insufficient range or unexpectedly negative numbers is left unaddressed,
as that would only happen in cases of journal corruption, so it is
likely that the problem directory would contain a bunch of garbage
anyway.