applet: application: Use GPtrArray
Having a linked list as a “deferred queue” is inherently dangerous due
to some functions removing elements from the list and not being able to
update the (potentially) changed head. Moreover, linked lists are the
leading cause of memory fragmentation in the world (even though we
usually don’t have that many objects flying around). This commit changes
the global list of deferred problems to a field in AbrtAppletApplication
and updates all callers accordingly.