daemon: Refactor signal pipe callback
The body of the inner loop in handle_signal_pipe_cb() could only ever be
run at most once because of the early return, thus only checking at most
one of the received signals.

The function was introduced in 99453e6e2 and the original intent of why
the buffer has size 2 or what to do if two signals arrive in the pipe is
not clear to me. Hence I removed the inner loop to get a functionally
equivalent code.

Issue reported by Coverity.