oops-utils: Respect the 'world-readable' flag
Originally, the only effect of the ABRT_OOPS_WORLD_READABLE flag was to
store the owner in the metadata directory as 'nobody'.

This commit reflects the flag in the filesystem permissions as well
by allowing world to read the directory or, in chmod symbolic language,
setting the mode to o+r.

Resolves #1515