Add debuginfod support to abrt-retrace-client
Debuginfod is an HTTP server for distributing ELF, DWARF and source
code.

This patch adds a '--debuginfod' option to abrt-retrace-client which
causes a 'X-Debuginfod' header to be added to the request.

This header will indicate to retrace-server that it shouldn't attempt
to aquire any packages needed to produce a backtrace (assuming retrace-server
is configured to allow debuginfod support).

Instead GDB's built-in debuginfod client will query debuginfod servers for any
missing debugging resources when retrace-server invokes it to generate a
backtrace from a corefile.

This patch also updates documentation to mention debuginfod support.

For more information regarding debuginfod, see
https://sourceware.org/elfutils/Debuginfod.html