abrt-dump-journal: Don't die on corrupt data
abrt-dump-journal-xorg and abrt-dump-journal-oops die when they
encounter a corrupt journal file. This makes restarting their
respective daemons impossible because the program crashes before it has
a chance to set and save its journal read cursor past the corrupt data.

This commit makes the programs go on with just a warning so there's an
opportunity to save the read cursor and thus recover from the condition.

This only affects abrt-dump-journal-xorg and abrt-dump-journal-oops when
run in the follow/daemon mode (i.e. with the -f option).

Resolves:
https://bugzilla.redhat.com/show_bug.cgi?id=1433652
https://bugzilla.redhat.com/show_bug.cgi?id=2043495