a-dump-journal-core: First seek the journal tail and then set filters
See: rhbz#2128662

a-dump-journal-core uses filters to jump between
systemd-coredump entries in journal.

However, we should first jump to the starting position
(journal tail) and only then set filters.
I suspect that setting the filters early, before we are
at the starting position, can lead to the strange
behavior that we are seeing in Fedora.

This patch makes the problem go away on my freshly-installed
Fedora 37 VM.

Signed-off-by: Michal Srb <michal@redhat.com>