Replace now gone pgpHexStr() with rpmhex()
In RPM 4.19, the pgpHexStr() function was renamed to rpmhex(), much like
the comment here in the code suggests.  The function was deprecated in
RPM 4.18 while keeping the original shim.  That's no longer the case in
4.19, though, where it was removed.

Now that Fedora 38 is the latest stable release and Fedora 39 is getting
RPM 4.19, it's probably time to finally make the move.