applet: only show the restart button for apps that are listed in menus
If the application is not listed in app menus, i.e. it has
NoDisplay=true in its .desktop file, then it probably means that
it is not an application that users typically start directly.
For example gnome-shell would fit this criteria.

We won't show the restart button for such applications.

Signed-off-by: Michal Srb <michal@redhat.com>