ccpp: don't fail bug reporting on ureport failure
Related: gh#1656

ureport works with the core backtrace retrieved from coredumpctl,
but the backtrace is sometimes quite useless:

    Stack trace of thread 3219:
    #0  0x0000aaaaca9d02dc crash.constprop.0 (will_segfault + 0x102dc)
    #1  0x0070aaaaca9d0370 n/a (n/a + 0x0)
    ...
    #9  0x0013ffff8dd40abc n/a (n/a + 0x0)
    #10 0x0076aaaaca9d0170 n/a (n/a + 0x0)
    ELF object binary architecture: AARCH64

abrt (understadably) complains about such low value backtrace and refuses
to continue. However, when we later run gdb on the coredump,
we actually get a good backtrace out of it.

This commit makes ureport failures non-blocking and thus there is a chance
that we will get a good quality backtrace from gdb later.

Signed-off-by: Michal Srb <michal@redhat.com>