cli: use pkgutil to discover modules
For a sourceless Python environment, it is valid to have only a .pyc
bytecode file in place of .py source. This is useful for saving space
while maintaining performance on an embedded system.

Rather than alter the logic to search for both .py and .pyc, use
Python's own logic to enumerate the available modules in the current
directory. This is simpler and should be correct for all cases that
Python supports.

Signed-off-by: Tyler Hall <tylerwhall@gmail.com>